+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/screen/VpnSettingsScreen.kt
@@ -926,20 +926,14 @@ fun VpnSettingsContent(
                         SelectableCell(
                             title =
                                 when (it.quantumResistantState) {
-                                    QuantumResistantState.Auto ->
-                                        stringResource(id = R.string.automatic)
-
                                     QuantumResistantState.Off -> stringResource(id = R.string.off)
-
                                     QuantumResistantState.On -> stringResource(id = R.string.on)
                                 },
                             isSelected = it.selected,
                             modifier =
                                 Modifier.animateItem()
                                     .then(
-                                        if (
-                                            it.quantumResistantState == QuantumResistantState.Auto
-                                        ) {
+                                        if (it.quantumResistantState == QuantumResistantState.On) {
                                             Modifier.focusRequester(
                                                 focusRequesters.getValue(
                                                     FeatureIndicator.QUANTUM_RESISTANCE
@@ -954,7 +948,6 @@ fun VpnSettingsContent(
                             },
                             testTag =
                                 when (it.quantumResistantState) {
-                                    QuantumResistantState.Auto -> ""
                                     QuantumResistantState.On -> LAZY_LIST_QUANTUM_ITEM_ON_TEST_TAG
 
                                     QuantumResistantState.Off -> LAZY_LIST_QUANTUM_ITEM_OFF_TEST_TAG