+++ mullvad-relay-selector/src/relay_selector/query.rs
@@ -206,7 +206,7 @@ pub struct WireguardRelayQuery {
     pub obfuscation: ObfuscationQuery,
     pub daita: Constraint<bool>,
     pub daita_use_multihop_if_necessary: Constraint<bool>,
-    pub quantum_resistant: QuantumResistantState,
+    pub quantum_resistant: Constraint<QuantumResistantState>,
 }
 
 #[derive(Default, Debug, Clone, Eq, PartialEq)]
@@ -295,7 +295,7 @@ impl WireguardRelayQuery {
 }
 
 impl WireguardRelayQuery {
-    pub const fn new() -> WireguardRelayQuery {
+    pub fn new() -> WireguardRelayQuery {
         WireguardRelayQuery {
             port: Constraint::Any,
             ip_version: Constraint::Any,
@@ -307,7 +307,7 @@ impl WireguardRelayQuery {
             obfuscation: ObfuscationQuery::Auto,
             daita: Constraint::Any,
             daita_use_multihop_if_necessary: Constraint::Any,
-            quantum_resistant: QuantumResistantState::Auto,
+            quantum_resistant: Constraint::Any,
         }
     }
 
@@ -575,7 +575,7 @@ pub mod builder {
         pub fn quantum_resistant(
             mut self,
         ) -> RelayQueryBuilder<Wireguard<Multihop, Obfuscation, Daita, bool>> {
-            self.query.wireguard_constraints.quantum_resistant = QuantumResistantState::On;
+            self.query.wireguard_constraints.quantum_resistant = QuantumResistantState::On.into();
             // Update the type state
             RelayQueryBuilder {
                 query: self.query,