+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Line wrap the file at 100 chars.                                              Th
 ### Removed
 - Remove OpenVPN support from the app.
   - See blog post for more info: https://mullvad.net/en/blog/removing-openvpn-15th-january-2026
+- Remove "Automatic" as a setting for the "Quantum-resistant tunnel" option.
 
 
 ## [2025.13-beta2] - 2025-10-28
@@ -53,6 +54,7 @@ Line wrap the file at 100 chars.                                              Th
 - Update Electron from 36.5.0 to 37.6.0.
 - Run version check hourly and when interacting with the app instead of once per day.
 - Add support for gradual rollouts of new app releases
+
 
 ### Fixed
 #### macOS